backend/requirements.txt
fastapi==0.115.0
httpx==0.28.1
joblib==1.5.3
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.5.0
pydantic_core==2.41.5
python-dotenv==1.2.1
scikit-learn==1.5.0
shap==0.50.0
supabase==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
uvicorn==0.30.0
xgboost==3.1.3
